requirements.txt
Flask==1.1.2
flask-restful==0.3.8
Flask-Cors==3.0.8
markdown==3.2.2
sqlalchemy==1.3.19
PyNaCl==1.4.0
Werkzeug==0.15.5
twilio~=6.50.1
sendgrid==6.9.7
